refactor: chunk-review 스크립트 경로 이동 및 상수 수정
