+++ scripts/chunk-review/worker.ts
@@ -3,7 +3,7 @@
 import fs from 'fs';
 import path from 'path';
 
-import { initLogger, logInfo, logError, logStart, logEnd } from '../../../lib/logging';
+import { initLogger, logInfo, logError, logStart, logEnd } from '../../lib/logging';
 import {
   splitCommand,
   atomicWriteJsonAsync,
@@ -13,9 +13,9 @@ import {
   runWithRetry as sharedRunWithRetry,
   MAX_RETRIES,
   BASE_DELAY_MS,
-} from '../../../lib/worker-utils';
+} from '../../lib/worker-utils';
 
-const PROMPTS_DIR = path.resolve(import.meta.dirname, '../prompts');
+const PROMPTS_DIR = path.join(path.resolve(import.meta.dirname, '../..'), 'skills', 'code-review', 'prompts');
 const FALLBACK_FILE = 'reviewer.md';
 
 // ---------------------------------------------------------------------------
@@ -91,7 +91,7 @@ function main() {
   const timeoutSec = options.timeout ? Number(options.timeout) : 0;
 
   // Initialize persistent logging
-  const projectRoot = path.resolve(import.meta.dirname, '../../..');
+  const projectRoot = path.resolve(import.meta.dirname, '../..');
   const jobId = jobDir ? path.basename(jobDir).replace(/^chunk-review-/, '') : 'unknown';
   initLogger('chunk-review-worker', projectRoot, jobId);
   logStart();
